put onto computer

+++ Physics Simulator/Assets/MyMaths.cs
@@ -2,21 +2,20 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class MyMaths : MonoBehaviour {
+public class MyMaths : MonoBehaviour
+{
 
     public static float Magnitude(float value)
     {
         if (value < 0)
         {
             return -value;
-    
-    }
+        }
         else
         {
             return value;
         }
     }
-
     public static float SquareRoot(float x)
     {
         if (x <= 0)
@@ -33,7 +32,6 @@ public class MyMaths : MonoBehaviour {
         }
         return current;
     }
-
     public static float Clamp(float value, float min, float max)
     {
         if (min > max)
@@ -52,7 +50,6 @@ public class MyMaths : MonoBehaviour {
         }
         return value;
     }
-
     public static float Vector_Magnitude(List<float> a)
     {
         float SquaredMagnitude = 0;
@@ -62,7 +59,6 @@ public class MyMaths : MonoBehaviour {
         }
         return SquareRoot(SquaredMagnitude);
     }
-
     public static float DotProduct_Value(List<float> a, List<float> b)
     {
         if (a.Count != b.Count)
@@ -79,7 +75,6 @@ public class MyMaths : MonoBehaviour {
             return counter;
         }
     }
-
     public static float DotProduct_Angle(List<float> a, List<float> b)
     {
         if (a.Count != b.Count)
@@ -88,13 +83,9 @@ public class MyMaths : MonoBehaviour {
         }
         else
         {
-            float Sinangle = (float)Vector_Magnitude(a) * Vector_Magnitude(b) / DotProduct_Value(a, b);
+            float Sinangle = Vector_Magnitude(a) * Vector_Magnitude(b) / DotProduct_Value(a, b);
             //arcsin means sin^-1 so inverse of sin
             return Mathf.Asin(Sinangle);
         }
     }
-
-
-
-
 }
